feat: 자동차 이름에 한글, 영어 이외의 문자도 허용하도록 기능 변경
(cherry picked from commit 6c1a64ae688a9e056266a95ece2c2aa433c91cb9)
