+++ README.md
@@ -3,14 +3,12 @@
 - [x] 사용자가 잘못된 입력을 할 경우에는 그 부분부터 재입력을 받는다.
 
 - 경주할 자동차 이름을 입력받는다.
-    - [x] 자동차 이름은 쉼표(,)로만 구분 가능하다.
-    - [x] 자동차 이름을 입력받을 때 쉼표로 끝날 수 없다.
+    - [x] 자동차 이름은 쉼표(,)로만 구분되어야 한다.
     - [x] 2대 이상의 자동차가 참여해야 한다.
-    - [x] 이름은 5자 이하만 가능하다.
-    - [x] 자동차 이름은 문자로만 입력 가능하다. (영어, 한글)
+    - [x] 자동차 이름은 5자 이하만 가능하다.
     - [x] 중복된 자동차 이름을 입력할 수 없다.
-    - [x] null 또는 빈 값을 입력할 수 없다.
-    - [x] 자동차 이름 앞 뒤에 공백이 존재해선 안된다.
+    - [x] 자동차 이름은 빈칸일 수 없다.
+    - [x] 자동차 이름 앞 뒤에 공백이 존재할 수 없다.
     - [x] 경주에 참여 가능한 자동차는 최대 20대까지이다.
 
 - 시도할 횟수를 입력받는다.
